Start parsing filtering expressions.
Add a new class FilterAWSCompiler that inherits from ASTVisitor to
parse the filtering expression. The Adapter supports a list of
product= or item= terms combined with OR that can be used to select
the set of products or items for which to return information. It
returns a list of conditions/items that will be used by the read
adapter to control what is monitored. (The conditions aren't actually
used yet, that will be in the next commit).

Add an option -cloudwatch true that controls whether or not the
returned data includes CloudWatch metrics. Default is true.

Add a set of unit tests that test positive and negative cases for
parsing the filtering expression.

Fill in the Read Options section of the README discussing the
filtering expressions and how they work.

Also fix one typo where I missed an "integration" -> "adapter" rename.