Finish supporting filter expressions.
Make some changes to further isolate the read proc from the details of
the AWS plugins. A new file aws/Factory.js contains the set of modules
and has functions to initialize the AWS module, create a plugin from a
product name, return the list of supported product names, etc. This
simplifies the read proc a bit.

Modify every aws module to support both class static functions as well
as instance methods to return the aws product name associated with a
module as well as the item name "InstanceId, VolumeId, etc."
associated with a given module.

Clean up the instance variables to the read proc to be more consistent
about using _<variable> for private variables.

Add a new CloudWatch method monitor_item which tells CloudWatch to
monitor a specific item rather than all items for a given
product. This updates a single list of products to monitor which
either has an empty list of items (monitor all items for a given
product) or a list of specific items.

The products array from the filter expression map to
CloudWatch::monitor_product() calls, and the items array from the
filter expression maps to CloudWatch::monitor_item() calls.