Remove CloudWatch half of adapter.
Remove all the CloudWatch capabilities of the adapter, which will be
handled by a separate juttle-cloudwatch-adapter. This includes
filtering on items, a -cloudwatch option to the read command, the
CloudWatch plugin, etc.

What remains is demographic/aggregate/metric/event information
obtained by the non-cloudwatch api calls and driven by plugins.

I'll have some additional changes to tighten the adapter contract to
support live-only reads but that's waiting on a final version of the
api.