Update to latest official read api.
Update to final released version of read api. The AWS adapter sets
periodicLiveRead to true so it will be periodically called when run
with -every. The default values for from and to are from=now,
to=infinity.

Also, the aws api only supports reading live (current) information, so
throw a RT-ADAPTER-UNSUPPORTED-TIME-OPTION error if -from is set to
the past.

The plugin poll() methods were aleady set up to only take a now value
to put in datapoints, so just use the value of from for that.