Aggregate in juttle vs in adapter.
Perform aggregations/demographics entirely in Juttle
instead of in the adapter.

A new file aws_module.juttle defines subs that perform the
aggregations/demographic calculations that are currently done by the
adapter code. The subs perform batching every second. This should work
regardless of the value chosen for -every, as all points that come out
of a single call to poll() have the same timestamp (namely, the from
value to the adapter read). To support events, add a path that only
takes points with event_type and does nothing. The aggregations happen
for all points without an event_type field.

To aggregate the raw points, use the module as follows:

-----
import "aws_module.juttle" as AWSMod;

read aws -from :now: -to :now: product="RDS" | AWSMod.aggregate_all | viww table
-----

The read command now returns the raw json points from the various
product AWS APIs, and the aggregations summarize them down to the same
points that used to come out of the adapters directly.

Remove Demographic.js, as it's not needed any longer.

Move the differencing for sets of items into the base class. Its
poll() method now calls get_items() to fetch the current set of items,
as a hash item name -> item json. It does the work of looping over the
item jsons to tidy each and build a list of points. It also does the
set differencing to compare the current and prior set of items and
emit points for any differences.

This requires a fix for juttle/juttle#351, as there are several places
where reducing can result on empty batches of points, resulting in
time values that are null.

Also update the README to reflect that by default the output points
are the raw results of the AWS API calls, but can be aggregated by
using the aws module. Intentionally vague on how the module can be
imported as we need to see how juttle/juttle#325 shakes out.