Address feedback from PR.
- Remove trailing whitespace from all files and extra whitespace before loops.
- Change pure virtual methods product()/item_name()/get_items() to throw if not overridden.
- Have AWSMon.changes handle the initial case where there are no existing items.
- Simplify the logic in changes for the case where nothing has changed.
- Fix some incorrect checks for next token in paginations.
- Instead of catching errors in each plugin's get_items() method,
  catch it once in AWSMon's poll() method, which just logs an error
  and re-throws.
- Make sure log levels are consistent across plugins.
- For the functions that return promises recursively, just return
  nothing for the base case instead of a resolved promise.
- Rely more heavily on ES6, making the set of products a CONST, the Factory a class with static methods, etc.
- Lambda has a function ending in Async, which normally prevents
  promisifyAll from working. Instead, use the suffix Bsync with
  promisifyAll.
- When parsing the filtering expression, use a single method
  throwUnsupportedFilter to save some duplicated code.
- When loading the adapter, make sure the config has the required
  access_key/secret_key/region properties.
- Don't need to check for allowed options or timeranges any longer,
  those are handled by the base class.
- Use Promise.map instead of Promise.all on an array of promises.

Add better unit tests that test the two cases of raw info fetch and
aggregations through the aws_module.juttle file.

The raw fetches don't test much other than the time and product
existing (matching one of the supported products) as well as an item
id with the correct name.

The aggregate validations are a bit more sophisticated, checking for
all supported fields and making sure that the aggregates and
demographics for each product are in the set of expected values for
that product.

Found a couple of typos in the docs while doing this.

Also fix lint.