Update for Juttle 0.5.0.
Access AdapterRead, JuttleMoment, StaticFilterCompiler etc through
JuttleAdapterAPI.

Update defaultTimeRange to defaultTimeOptions.

Use ES6 class for FilterAWSCompiler derived from StaticFilterCompiler.

Use already-provided featureNotSupported to return errors.

Update filter parsing callbacks based on the advice in
https://github.com/juttle/juttle/wiki/Porting-Adapters-to-the-0.5.x-API.

Create the adapter in the test as suggested.

Fix a false negative problem in the tests where one of the programs
that should have returned an error was falsely succeeding (-from
before :now:). And then add a check in the constructor to fix that.

-from :now: -to :now: isn't supported any longer, so replace it with
 -from :now: -to :1 second from now: to get a single read.

Note the juttleAdapterAPI dependency in package.json.