Change default timerange.
Instead of having the default timerange be from :now: to :end:, change
the default to be from :now: - epsilon to :now. This does change when
someone has to specify a -from/-to for the two common use cases of a
one-shot read vs reading live indefinitely, but it's easier to just
add a -to :end: than add a -to :1 second from now:, so that's a bit
more compact.